evaluation/pyproject.toml
[tool.poetry]
name = "hackathon-dev"
version = "0.1.0"
description = ""
authors = ["Greg Chase <9953173+gregwchase@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
openai = "^1.24.0"
arize-phoenix = {extras = ["evals"], version = "^3.24.0"}
arize = {extras = ["tracing"], version = "^7.16.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
